:recycle: code refactor
