+++ classicalBloom/bitarray.js
